+++ app/src/androidTest/java/com/swent/assos/NewsTest.kt
@@ -1,7 +1,6 @@
 package com.swent.assos
 
 import androidx.activity.compose.setContent
-import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.kaspersky.components.composesupport.config.withComposeSupport
@@ -12,7 +11,7 @@ import com.swent.assos.model.navigation.NavigationActions
 import com.swent.assos.screens.AssoDigestScreen
 import com.swent.assos.screens.CreateNewsScreen
 import com.swent.assos.screens.NewsScreen
-import com.swent.assos.ui.screens.AssoDigest
+import com.swent.assos.ui.screens.manageAssos.AssoDigest
 import com.swent.assos.ui.screens.News
 import com.swent.assos.ui.screens.manageAssos.CreateNews
 import dagger.hilt.android.testing.HiltAndroidRule
